db-flask-backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --extra-index-url='https://oauth2accesstoken:****@us-central1-python.pkg.dev/releaseinf/oanda-python-repo/simple/' --no-emit-index-url --no-emit-trusted-host --output-file=requirements.txt requirements.in
#
alembic==1.7.7
    # via flask-migrate
apispec==5.1.1
    # via -r requirements.in
atlassian-python-api==3.32.2
    # via -r requirements.in
attrs==21.4.0
    # via jsonschema
cachecontrol==0.12.10
    # via -r requirements.in
cachetools==4.2.4
    # via google-auth
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via flask
cryptography==37.0.1
    # via -r requirements.in
dependency-injector==4.37.0
    # via -r requirements.in
deprecated==1.2.13
    # via atlassian-python-api
flasgger==0.9.5
    # via -r requirements.in
flask==2.0.2
    # via
    #   -r requirements.in
    #   flasgger
    #   flask-cors
    #   flask-migrate
    #   flask-sqlalchemy
    #   prometheus-flask-exporter
flask-cors==3.0.10
    # via -r requirements.in
flask-migrate==3.1.0
    # via -r requirements.in
flask-sqlalchemy==2.5.1
    # via
    #   -r requirements.in
    #   flask-migrate
google-auth==2.3.3
    # via
    #   -r requirements.in
    #   google-auth-oauthlib
google-auth-oauthlib==0.4.6
    # via -r requirements.in
greenlet==2.0.2
    # via sqlalchemy
gunicorn==20.1.0
    # via -r requirements.in
idna==3.3
    # via requests
itsdangerous==2.1.2
    # via flask
jinja2==3.1.2
    # via flask
jsonschema==4.4.0
    # via flasgger
mako==1.2.0
    # via alembic
markupsafe==2.1.1
    # via
    #   jinja2
    #   mako
marshmallow==3.14.1
    # via -r requirements.in
mistune==2.0.2
    # via flasgger
msgpack==1.0.3
    # via cachecontrol
mysql-connector-python==8.0.27
    # via -r requirements.in
oauthlib==3.2.0
    # via
    #   atlassian-python-api
    #   requests-oauthlib
prometheus-client==0.12.0
    # via
    #   -r requirements.in
    #   prometheus-flask-exporter
prometheus-flask-exporter==0.18.6
    # via -r requirements.in
protobuf==3.20.1
    # via mysql-connector-python
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via -r requirements.in
pytz==2021.3
    # via -r requirements.in
pyyaml==6.0
    # via flasgger
requests==2.26.0
    # via
    #   -r requirements.in
    #   atlassian-python-api
    #   cachecontrol
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via
    #   atlassian-python-api
    #   google-auth-oauthlib
rsa==4.8
    # via google-auth
six==1.16.0
    # via
    #   atlassian-python-api
    #   dependency-injector
    #   flasgger
    #   flask-cors
    #   google-auth
    #   python-dateutil
sqlalchemy==1.4.36
    # via
    #   alembic
    #   flask-sqlalchemy
urllib3==1.26.9
    # via requests
werkzeug==2.1.2
    # via flask
wrapt==1.15.0
    # via deprecated

# The following packages are considered to be unsafe in a requirements file:
# setuptools
